+++ packages/contracts/src/index.ts
@@ -2,3 +2,4 @@ export * from "./artifact.js";
 export * from "./event.js";
 export * from "./ids.js";
 export * from "./task.js";
+export * from "./workspace.js";